Changed files and moved them to one directory instead of many separate ones. As well as merged logic from http.c to install.c
