+++ src/CMakeLists.txt
@@ -2,10 +2,5 @@ cmake_minimum_required(VERSION 3.22.0)
 
 project(japm)
 
-add_subdirectory(Errors)
-add_subdirectory(PackageManagment)
-
-add_executable(${PROJECT_NAME} main.c)
-
-target_link_libraries(${PROJECT_NAME} errors)
-target_link_libraries(${PROJECT_NAME} install_package)
+add_executable(${PROJECT_NAME} main.c errors.c install.c)
+target_link_libraries(${PROJECT_NAME} json-c)
