--- src/Errors/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_library(errors errors.c errors.h)
