--- src/PackageManagment/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_library(install_package install.c install.h)
-target_link_libraries(install_package -ljson-c HTTP)
-add_subdirectory(HTTP)
