--- src/PackageManagment/HTTP/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_library(HTTP http.c http.h)
