--- src/PackageManagment/HTTP/http.h
@@ -1,2 +1,0 @@
-int http_req(char*url);
-void download_package(char*url, char*package_name);
