--- src/PackageManagment/program.json
@@ -1,5 +1,0 @@
-{
-    "name": "My Package",
-    "dependencies": [],
-    "install" : "ls"
-}
