+++ src/errors.c
