+++ src/errors.h
