+++ src/install.c
@@ -3,8 +3,7 @@
 #include <string.h>
 #include <json-c/json.h>
 
-#include "../Errors/errors.h"
-#include "HTTP/http.h"
+#include "errors.h"
 
 // Some hardcoded values that are used in the program
 #define MAX_PACKAGE_FILE_SIZE 10240
@@ -16,6 +15,8 @@ struct package_information parse_package_information(char *package_name);
 void install_package_to_system(struct package_information package_info);
 int check_if_package_installed(char *package_name);
 void add_package_to_installed_packages(char *package_name);
+int http_req(char *url);
+void download_package(char *url, char *package_name);
 
 struct package_information
 {
@@ -237,3 +238,50 @@ check_if_package_installed(char* package_name)
         return 0;
     }
 }
+
+int http_req(char*url) 
+{
+    /* Create a HTTP request to the provided argument url and return the response code */
+    /* To do the request I will call curl command using the system function */
+    /* The response code would be saved in a temporary file that would then be read */
+    /* The response code would be returned */
+    char *command = malloc(sizeof(char) * (strlen("curl -s -o /dev/null -w %{http_code} ") + strlen(url) + strlen(" > /tmp/reponse_code.txt") + 1));
+    strcpy(command, "curl -s -o /dev/null -w %{http_code} ");
+    strcat(command, url);
+    strcat(command, " > /tmp/response_code.txt");
+    system(command);
+
+    FILE *fp = fopen("/tmp/response_code.txt", "r");
+    if (fp == NULL)
+    {
+        printf("Something went wrong. Check your internet connection\n");
+        exit(unkown_error);
+    }
+
+    // Read the response code from the file
+    char response_code[4];
+    fscanf(fp, "%s", response_code);
+    fclose(fp);
+    free(command);
+
+    //Convert the reponse code to integer and return it
+    return atoi(response_code);
+}
+
+void 
+download_package(char*url, char*package_name)
+{
+    // Download the package file from the provided url
+    // The file would be saved in /var/cache/japm/package_name
+    
+    system("mkdir -p /var/cache/japm"); // The cache directory for japm might not exist so we need to first create it using the mkdir command
+
+    char *command = malloc(sizeof(char) * (strlen(url) + strlen("curl -s -o /var/cache/japm/") + strlen(package_name) + 1));
+    strcpy(command, "curl -s -o /var/cache/japm/");
+    strcat(command, package_name);
+    strcat(command, " ");
+    strcat(command, url);
+    system(command);
+    free(command);
+}
+
